requirements.txt
# Core
numpy>=1.26
pandas>=2.2
scipy>=1.13
matplotlib>=3.8
seaborn>=0.13
plotly>=5.23
rich>=13.7
tqdm>=4.66

# Apps / CLI
streamlit>=1.37
typer>=0.12

# Password hashing
bcrypt>=4.1
argon2-cffi>=23.1

# Length extension (pure Python)
hlextend>=0.1.2

# Dev / Notebook
jupyter>=1.0
notebook>=7.2
